+++ docs/sphinx/source/ReleaseNotes.md
@@ -7,6 +7,101 @@ As the [versioning guide](Versioning.md) details, it cannot always be determined
 
 ## 4.10
 
+### 4.10.6.0
+
+<h4> New Features </h4>
+
+* Add lucene queue size timer event - [PR #3988](https://github.com/FoundationDB/fdb-record-layer/pull/3988)
+
+
+**[Full Changelog (4.10.5.0...4.10.6.0)](https://github.com/FoundationDB/fdb-record-layer/compare/4.10.5.0...4.10.6.0)**
+
+#### Mixed Mode Test Results
+
+Mixed mode testing run against the following previous versions:
+
+✅`4.9.4.0`, ✅`4.9.6.0`, ✅`4.9.7.0`, ❌`4.9.8.0`, ✅`4.9.9.0`, ✅`4.10.1.0`, ✅`4.10.2.0`, ✅`4.10.3.0`, ✅`4.10.4.0`, ✅`4.10.5.0`
+
+[See full test run](https://github.com/FoundationDB/fdb-record-layer/actions/runs/22728009119)
+
+
+
+### 4.10.5.0
+
+<h4> New Features </h4>
+
+* Allow leaving corrupted indexes as readable after header repair - [PR #3801](https://github.com/FoundationDB/fdb-record-layer/pull/3801)
+<h4> Bug Fixes </h4>
+
+* Lucene: Convert  merge & drain pending queue to a future - [PR #3982](https://github.com/FoundationDB/fdb-record-layer/pull/3982)
+* Fix verify exception that could be forwarded if the `ORDER BY` expression was on a join constituent not in the final projection - [PR #3989](https://github.com/FoundationDB/fdb-record-layer/pull/3989)
+
+<details>
+<summary>
+
+<h4> Build/Test/Documentation/Style Improvements (click to expand) </h4>
+
+</summary>
+
+* Yaml tests with large records (split across multiple k/v) - [PR #3986](https://github.com/FoundationDB/fdb-record-layer/pull/3986)
+* Add `!randomStr` custom YAML tag for large string testing - [PR #3984](https://github.com/FoundationDB/fdb-record-layer/pull/3984)
+* Add an online indexing test for vector indexes - [PR #3972](https://github.com/FoundationDB/fdb-record-layer/pull/3972)
+
+</details>
+
+
+**[Full Changelog (4.10.4.0...4.10.5.0)](https://github.com/FoundationDB/fdb-record-layer/compare/4.10.4.0...4.10.5.0)**
+
+#### Mixed Mode Test Results
+
+Mixed mode testing run against the following previous versions:
+
+❌`4.9.3.0`, ✅`4.9.4.0`, ✅`4.9.6.0`, ✅`4.9.7.0`, ✅`4.9.8.0`, ✅`4.9.9.0`, ✅`4.10.1.0`, ✅`4.10.2.0`, ✅`4.10.3.0`, ✅`4.10.4.0`
+
+[See full test run](https://github.com/FoundationDB/fdb-record-layer/actions/runs/22710891170)
+
+
+
+### 4.10.4.0
+
+<h4> New Features </h4>
+
+* Add basic `ORDER BY` for `JOIN` queries with single leg orders - [PR #3867](https://github.com/FoundationDB/fdb-record-layer/pull/3867)
+* Include stats for each planner phase in EXPLAIN metrics - [PR #3512](https://github.com/FoundationDB/fdb-record-layer/pull/3512)
+* `INNER JOIN` syntax now supports `USING` - [PR #3893](https://github.com/FoundationDB/fdb-record-layer/pull/3893)
+<h4> Bug Fixes </h4>
+
+* Fix Incorrect Plan Constraints Generated for Join Predicates - [PR #3971](https://github.com/FoundationDB/fdb-record-layer/pull/3971)
+* refetch node references for the current layer - [PR #3966](https://github.com/FoundationDB/fdb-record-layer/pull/3966)
+
+<details>
+<summary>
+
+<h4> Build/Test/Documentation/Style Improvements (click to expand) </h4>
+
+</summary>
+
+* Parallelize the test steps of the release process - [PR #3977](https://github.com/FoundationDB/fdb-record-layer/pull/3977)
+* Lucene queue: add end to end serialization/deserialization tests - [PR #3963](https://github.com/FoundationDB/fdb-record-layer/pull/3963)
+* Parallelize mixed-mode testing - [PR #3964](https://github.com/FoundationDB/fdb-record-layer/pull/3964)
+* Lucene pending write queue: Add a concurrency test - [PR #3961](https://github.com/FoundationDB/fdb-record-layer/pull/3961)
+* `!pos` tag for `yaml` tests - [PR #3946](https://github.com/FoundationDB/fdb-record-layer/pull/3946)
+
+</details>
+
+
+**[Full Changelog (4.10.3.0...4.10.4.0)](https://github.com/FoundationDB/fdb-record-layer/compare/4.10.3.0...4.10.4.0)**
+
+#### Mixed Mode Test Results
+
+Mixed mode testing run against the following previous versions:
+
+❌`4.9.2.0`, ❌`4.9.3.0`, ✅`4.9.4.0`, ✅`4.9.6.0`, ✅`4.9.7.0`, ✅`4.9.8.0`, ✅`4.9.9.0`, ✅`4.10.1.0`, ✅`4.10.2.0`, ✅`4.10.3.0`
+
+[See full test run](https://github.com/FoundationDB/fdb-record-layer/actions/runs/22452643118)
+
+
+
 ### 4.10.3.0
 
 <h4> New Features </h4>